smm: Don't get stuck trying to accept a new search
If the smm server refuses to accept us starting a search indefinitely,
it would cause the main loop to get stuck and never progress.
This resulted in no progress being made on events which meant no more
updates were sent to flight safety or search map of position/status,
and no more commands would be actioned.